Add confirm pop-up before deleting analysis result.

+++ src/TextAnalysisComponents/ResultList.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { useState } from 'react';
 import { useEffect } from 'react';
 import { useHistory, withRouter } from 'react-router-dom';
-import {addAnalysisResult} from './AddAnalysisResult';
+import { addAnalysisResult } from './AddAnalysisResult';
 import Results from './Results';
 import Result from './Result'
 
@@ -11,7 +11,7 @@ const ResultList = () => {
 
     const [analysisResults, setAnalysisResult] = useState([])
 
-   
+
     useEffect(() => {
         const getResults = async () => {
             const resultsFromServer = await fetchResults()
@@ -20,7 +20,7 @@ const ResultList = () => {
 
         getResults()
     }, [])
-   
+
 
     //分析結果をサーバーから取得
     const fetchResults = async () => {
@@ -28,32 +28,37 @@ const ResultList = () => {
         const data = await res.json()
 
         return data
-     
-        
+
+
     }
 
-      // 分析結果の削除 
-  const deleteResult = async (id) => {
-      alert("削除しましょうか？")
-    await fetch(`http://localhost:5200/analysisResults/${id}`, {
-      method: 'DELETE',
-    })
+    // 分析結果の削除 
+    const deleteResult = async (id) => {
+        const isDelete = window.confirm("分析結果を削除しましょうか？")
+        if (isDelete) {
+            await fetch(`http://localhost:5200/analysisResults/${id}`, {
+                method: 'DELETE',
+            })
 
-    //  削除されたid以外のイベントのみを表示する
-    setAnalysisResult(analysisResults.filter((analysisResult) => analysisResult.id !== id))
+            //  削除されたid以外のイベントのみを表示する
+            setAnalysisResult(analysisResults.filter((analysisResult) => analysisResult.id !== id))
+        }
 
-  }
+        else {
+
+        }
+    }
 
     return (
-       <div>
-           <div className="kzHeader kzColor1 kzFont1">Kozipro</div>
-        
-        {/* <Results results={analysisResults}/> */}
-        {analysisResults.map((result,index)=>(
-             <Result key={index} result={result} onDelete={deleteResult}/>
-         ))}
-        
-       </div>
+        <div>
+            <div className="kzHeader kzColor1 kzFont1">Kozipro</div>
+
+            {/* <Results results={analysisResults}/> */}
+            {analysisResults.map((result, index) => (
+                <Result key={index} result={result} onDelete={deleteResult} />
+            ))}
+
+        </div>
     )
 }
